refactor(tests,layout): T8 shared SnapshotHarness + module moves (F41-F45)
