+++ crates/qedgen/src/project/validate.rs
