+++ crates/qedgen/tests/example_manifests.rs
@@ -1,12 +1,6 @@
-use std::path::{Path, PathBuf};
+mod common;
 
-fn repo_root() -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .parent()
-        .and_then(Path::parent)
-        .expect("qedgen crate should live under <repo>/crates/qedgen")
-        .to_path_buf()
-}
+use common::repo_root;
 
 #[test]
 fn rust_examples_with_qed_state_have_qed_toml() {
