+++ crates/qedgen/tests/sbpf_codegen_cli.rs
@@ -14,49 +14,18 @@
 //! emit, the Rust scaffold is suppressed, and the generated header
 //! names a command that actually exists.
 
+mod common;
+
+use common::{ensure_qedgen_built, qedgen_bin, repo_root};
 use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::process::Command;
-
-fn repo_root() -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .parent()
-        .and_then(Path::parent)
-        .expect("qedgen crate at <repo>/crates/qedgen")
-        .to_path_buf()
-}
-
-fn qedgen_bin() -> PathBuf {
-    let profile = if cfg!(debug_assertions) {
-        "debug"
-    } else {
-        "release"
-    };
-    repo_root().join("target").join(profile).join("qedgen")
-}
-
-fn ensure_qedgen_built() {
-    if !qedgen_bin().exists() {
-        let status = Command::new("cargo")
-            .args(["build", "--bin", "qedgen"])
-            .current_dir(repo_root())
-            .status()
-            .expect("spawn cargo build");
-        assert!(status.success(), "cargo build qedgen failed");
-    }
-}
 
 /// Fresh git-initialized tempdir containing the given spec file.
 fn sbpf_project(spec_src: &Path, spec_name: &str) -> tempfile::TempDir {
     let tmp = tempfile::tempdir().expect("create tempdir");
     fs::copy(spec_src, tmp.path().join(spec_name)).expect("copy spec fixture");
-    let git_init = Command::new("git")
-        .arg("init")
-        .arg("--quiet")
-        .current_dir(tmp.path())
-        .status()
-        .expect("spawn git init");
-    assert!(git_init.success(), "git init failed");
+    common::git_init(tmp.path());
     tmp
 }
